実装＿ランキングの反映
