+++ Assets/MyAssets/Develop/Moriwaki/Scenes/RankingTest.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 68272f7d0cefcb448a8eac395a2ad136
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
