+++ Assets/MyAssets/Develop/Moriwaki/Scripts/GetRanking.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e1c2313418760284295063c4453cbabb
+guid: 4b42eea8ecc31544685cca5fc268fcdb
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
