+++ ProjectSettings/TagManager.asset
@@ -11,7 +11,7 @@ TagManager:
   - 
   - Water
   - UI
-  - 
+  - Score
   - 
   - 
   -
